infra/assets/extra-requirements.txt
# Azure Functions core
azure-Functions>=1.24.0

# GitHub Copilot SDK
github-copilot-sdk==0.2.0

# Pydantic for tool parameter validation
pydantic>=2.0.0

# Parse AGENTS.md frontmatter and body
python-frontmatter>=1.1.0

azurefunctions-extensions-http-fastapi

# Azure connector tools (ARM API client)
aiohttp>=3.9.0
azure-identity>=1.15.0

# Azure Functions connector triggers
https://github.com/anthonychu/azure-functions-connectors-python/releases/download/v0.1.0b2/azure_functions_connectors-0.1.0b2-py3-none-any.whl